[FLINK-8555] [table] Simplify function operand checkers and inference
