+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/functions/utils/ScalarSqlFunction.scala
@@ -20,13 +20,12 @@ package org.apache.flink.table.functions.utils
 
 import org.apache.calcite.rel.`type`.RelDataType
 import org.apache.calcite.sql._
-import org.apache.calcite.sql.`type`.SqlOperandTypeChecker.Consistency
 import org.apache.calcite.sql.`type`._
 import org.apache.calcite.sql.parser.SqlParserPos
 import org.apache.flink.table.api.ValidationException
 import org.apache.flink.table.calcite.FlinkTypeFactory
 import org.apache.flink.table.functions.ScalarFunction
-import org.apache.flink.table.functions.utils.ScalarSqlFunction._
+import org.apache.flink.table.functions.utils.ScalarSqlFunction.createReturnTypeInference
 import org.apache.flink.table.functions.utils.UserDefinedFunctionUtils._
 
 import scala.collection.JavaConverters._
@@ -47,12 +46,12 @@ class ScalarSqlFunction(
   extends SqlFunction(
     new SqlIdentifier(name, SqlParserPos.ZERO),
     createReturnTypeInference(name, scalarFunction, typeFactory),
-    createOperandTypeInference(name, scalarFunction, typeFactory),
-    createOperandTypeChecker(name, scalarFunction),
+    createEvalOperandTypeInference(name, scalarFunction, typeFactory),
+    createEvalOperandTypeChecker(name, scalarFunction),
     null,
     SqlFunctionCategory.USER_DEFINED_FUNCTION) {
 
-  def getScalarFunction = scalarFunction
+  def getScalarFunction: ScalarFunction = scalarFunction
 
   override def isDeterministic: Boolean = scalarFunction.isDeterministic
 
@@ -91,113 +90,6 @@ object ScalarSqlFunction {
         val resultType = getResultTypeOfScalarFunction(scalarFunction, foundSignature.get)
         typeFactory.createTypeFromTypeInfo(resultType, isNullable = true)
       }
-    }
-  }
-
-  private[flink] def createOperandTypeInference(
-      name: String,
-      scalarFunction: ScalarFunction,
-      typeFactory: FlinkTypeFactory)
-    : SqlOperandTypeInference = {
-    /**
-      * Operand type inference based on [[ScalarFunction]] given information.
-      */
-    new SqlOperandTypeInference {
-      override def inferOperandTypes(
-          callBinding: SqlCallBinding,
-          returnType: RelDataType,
-          operandTypes: Array[RelDataType]): Unit = {
-
-        val operandTypeInfo = getOperandTypeInfo(callBinding)
-
-        val foundSignature = getEvalMethodSignature(scalarFunction, operandTypeInfo)
-          .getOrElse(
-            throw new ValidationException(
-              s"Given parameters of function '$name' do not match any signature. \n" +
-                s"Actual: ${signatureToString(operandTypeInfo)} \n" +
-                s"Expected: ${signaturesToString(scalarFunction, "eval")}"))
-
-        val inferredTypes = scalarFunction
-          .getParameterTypes(foundSignature)
-          .map(typeFactory.createTypeFromTypeInfo(_, isNullable = true))
-
-        for (i <- operandTypes.indices) {
-          if (i < inferredTypes.length - 1) {
-            operandTypes(i) = inferredTypes(i)
-          } else if (null != inferredTypes.last.getComponentType) {
-            // last argument is a collection, the array type
-            operandTypes(i) = inferredTypes.last.getComponentType
-          } else {
-            operandTypes(i) = inferredTypes.last
-          }
-        }
-      }
-    }
-  }
-
-  private[flink] def createOperandTypeChecker(
-      name: String,
-      scalarFunction: ScalarFunction)
-    : SqlOperandTypeChecker = {
-
-    val methods = checkAndExtractMethods(scalarFunction, "eval")
-
-    /**
-      * Operand type checker based on [[ScalarFunction]] given information.
-      */
-    new SqlOperandTypeChecker {
-      override def getAllowedSignatures(op: SqlOperator, opName: String): String = {
-        s"$opName[${signaturesToString(scalarFunction, "eval")}]"
-      }
-
-      override def getOperandCountRange: SqlOperandCountRange = {
-        var min = 254 // according to JVM spec 4.3.3
-        var max = -1
-        var isVarargs = false
-        methods.foreach(
-          m => {
-            var len = m.getParameterTypes.length
-            if (len > 0 && m.isVarArgs && m.getParameterTypes()(len - 1).isArray) {
-              isVarargs = true
-              len = len - 1
-            }
-            max = Math.max(len, max)
-            min = Math.min(len, min)
-          }
-        )
-        if (isVarargs) {
-          // if eval method is varargs, set max to -1 to skip length check in Calcite
-          max = -1
-        }
-        SqlOperandCountRanges.between(min, max)
-      }
-
-      override def checkOperandTypes(
-          callBinding: SqlCallBinding,
-          throwOnFailure: Boolean)
-        : Boolean = {
-        val operandTypeInfo = getOperandTypeInfo(callBinding)
-
-        val foundSignature = getEvalMethodSignature(scalarFunction, operandTypeInfo)
-
-        if (foundSignature.isEmpty) {
-          if (throwOnFailure) {
-            throw new ValidationException(
-              s"Given parameters of function '$name' do not match any signature. \n" +
-                s"Actual: ${signatureToString(operandTypeInfo)} \n" +
-                s"Expected: ${signaturesToString(scalarFunction, "eval")}")
-          } else {
-            false
-          }
-        } else {
-          true
-        }
-      }
-
-      override def isOptional(i: Int): Boolean = false
-
-      override def getConsistency: Consistency = Consistency.NONE
-
     }
   }
 }
