+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/functions/utils/TableSqlFunction.scala
@@ -18,19 +18,15 @@
 
 package org.apache.flink.table.functions.utils
 
-import org.apache.calcite.rel.`type`.RelDataType
 import org.apache.calcite.sql._
 import org.apache.calcite.sql.`type`._
 import org.apache.calcite.sql.parser.SqlParserPos
 import org.apache.calcite.sql.validate.SqlUserDefinedTableFunction
-import org.apache.calcite.sql.`type`.SqlOperandTypeChecker.Consistency
 import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.table.calcite.FlinkTypeFactory
 import org.apache.flink.table.functions.TableFunction
-import org.apache.flink.table.plan.schema.FlinkTableFunctionImpl
-import org.apache.flink.table.api.ValidationException
 import org.apache.flink.table.functions.utils.UserDefinedFunctionUtils._
-import org.apache.flink.table.functions.utils.TableSqlFunction._
+import org.apache.flink.table.plan.schema.FlinkTableFunctionImpl
 
 /**
   * Calcite wrapper for user-defined table functions.
@@ -45,8 +41,8 @@ class TableSqlFunction(
   extends SqlUserDefinedTableFunction(
     new SqlIdentifier(name, SqlParserPos.ZERO),
     ReturnTypes.CURSOR,
-    createOperandTypeInference(name, tableFunction, typeFactory),
-    createOperandTypeChecker(name, tableFunction),
+    createEvalOperandTypeInference(name, tableFunction, typeFactory),
+    createEvalOperandTypeChecker(name, tableFunction),
     null,
     functionImpl) {
 
@@ -69,113 +65,4 @@ class TableSqlFunction(
   override def isDeterministic: Boolean = tableFunction.isDeterministic
 
   override def toString: String = displayName
-}
-
-object TableSqlFunction {
-
-  private[flink] def createOperandTypeInference(
-    name: String,
-    tableFunction: TableFunction[_],
-    typeFactory: FlinkTypeFactory)
-  : SqlOperandTypeInference = {
-
-    /**
-      * Operand type inference based on [[TableFunction]] given information.
-      */
-    new SqlOperandTypeInference {
-      override def inferOperandTypes(
-          callBinding: SqlCallBinding,
-          returnType: RelDataType,
-          operandTypes: Array[RelDataType]): Unit = {
-
-        val operandTypeInfo = getOperandTypeInfo(callBinding)
-
-        val foundSignature = getEvalMethodSignature(tableFunction, operandTypeInfo)
-          .getOrElse(throw new ValidationException(
-            s"Given parameters of function '$name' do not match any signature. \n" +
-              s"Actual: ${signatureToString(operandTypeInfo)} \n" +
-              s"Expected: ${signaturesToString(tableFunction, "eval")}"))
-
-        val inferredTypes = tableFunction
-          .getParameterTypes(foundSignature)
-          .map(typeFactory.createTypeFromTypeInfo(_, isNullable = true))
-
-        for (i <- operandTypes.indices) {
-          if (i < inferredTypes.length - 1) {
-            operandTypes(i) = inferredTypes(i)
-          } else if (null != inferredTypes.last.getComponentType) {
-            // last argument is a collection, the array type
-            operandTypes(i) = inferredTypes.last.getComponentType
-          } else {
-            operandTypes(i) = inferredTypes.last
-          }
-        }
-      }
-    }
-  }
-
-  private[flink] def createOperandTypeChecker(
-    name: String,
-    tableFunction: TableFunction[_])
-  : SqlOperandTypeChecker = {
-
-    val methods = checkAndExtractMethods(tableFunction, "eval")
-
-    /**
-      * Operand type checker based on [[TableFunction]] given information.
-      */
-    new SqlOperandTypeChecker {
-      override def getAllowedSignatures(op: SqlOperator, opName: String): String = {
-        s"$opName[${signaturesToString(tableFunction, "eval")}]"
-      }
-
-      override def getOperandCountRange: SqlOperandCountRange = {
-        var min = 254
-        var max = -1
-        var isVarargs = false
-        methods.foreach( m => {
-          var len = m.getParameterTypes.length
-          if (len > 0 && m.isVarArgs && m.getParameterTypes()(len - 1).isArray) {
-            isVarargs = true
-            len = len - 1
-          }
-          max = Math.max(len, max)
-          min = Math.min(len, min)
-        })
-        if (isVarargs) {
-          // if eval method is varargs, set max to -1 to skip length check in Calcite
-          max = -1
-        }
-
-        SqlOperandCountRanges.between(min, max)
-      }
-
-      override def checkOperandTypes(
-        callBinding: SqlCallBinding,
-        throwOnFailure: Boolean)
-      : Boolean = {
-        val operandTypeInfo = getOperandTypeInfo(callBinding)
-
-        val foundSignature = getEvalMethodSignature(tableFunction, operandTypeInfo)
-
-        if (foundSignature.isEmpty) {
-          if (throwOnFailure) {
-            throw new ValidationException(
-              s"Given parameters of function '$name' do not match any signature. \n" +
-                s"Actual: ${signatureToString(operandTypeInfo)} \n" +
-                s"Expected: ${signaturesToString(tableFunction, "eval")}")
-          } else {
-            false
-          }
-        } else {
-          true
-        }
-      }
-
-      override def isOptional(i: Int): Boolean = false
-
-      override def getConsistency: Consistency = Consistency.NONE
-
-    }
-  }
 }
